Add plural aliases for organization endpoints
In most of the API, we've standardized on using plural nouns in endpoint
names, but `OrganizationController` predated that decision. Add
`/api/v1/organizations` as an alternate name for `/api/v1/organization`.
Once clients have moved over to the new name, we can get rid of the old one.

No functional change here. If clients are using the operation ID from the
OpenAPI schema to look up endpoint paths, they will switch over to the new
names automatically; otherwise the old names continue to work for now.