Remove SearchTable.leftJoinWithMain()
There is no longer a notion of a "main table" since searches can start at any
point in the hierarchy, so this was on outdated holdover from accession-centric
searching. Join conditions live in the sublist field definitions now.

This lets us get rid of AccessionParentTable and simplify the flattened
sublist join logic in both SearchService and NestedQueryBuilder.