Reduce noise in database tests (#112)
The Spring test framework logs a few very verbose INFO messages for each test
method as it sets up and tears down the database connection. This makes it more
of a pain to see the test results when running tests interactively.

Change the logging configuration to only show Spring log messages at WARN and
ERROR levels.