TRACK-464 Model sublist fields explicitly (#107)
This change introduces the `SublistField` class. As its name suggests, it
represents a sublist field, in the same way the existing `SearchField` class
represents a scalar field. This is a simpler and hopefully easier to understand
model of the search field hierarchy.

The new class, to be more precise, represents the _relationship_ between two
namespaces. Knowledge of whether a relationship is one-to-many or many-to-one now
lives in `SublistField`, which means there's no longer a need for separate
`SearchFieldNamespace` properties for single-value and multi-value sublists.

The SQL snippet required to connect a multiset query to its enclosing query also
lives in `SublistField` now. This replaces the `conditionForMultiset()` method on
`SearchTable`, which we need to do in order to fully support navigating
to tables such as `photos` that are referenced from multiple places in the data
model.

There are no behavior changes here; existing tests continue to pass as-is.