Don't include empty maps in search results
This fixes a bug that was masked by the fact that accession searches always
include the accession ID and accession number in the field list. If you
searched for a sublist field, you would get back an entry in the results list
for every root-level entity that matched the filter criteria, even if there
were no values for the sublist you were asking for. That violates the contract
of the search API, which says you will never get back a completely empty map.

Fix the bug and add a regression test to verify the fix.