TRACK-513 Add aliases for fields that ought to be sublists
There are a few accession fields that refer to values in secondary tables and
whose field names are the names that we'd ideally want to use for the
corresponding sublists. To avoid name collisions, the sublists have `Info`
suffixes for now.

Add alternate aliases the client can use temporarily. The migration sequence
will look like this:

1. Add alternate names (this change)
2. Update the client to use the alternate names
3. Remove the offending fields and rename the sublists
4. Update the client to use the flattened sublist names instead of the aliases