TRACK-485 Add namespaces for accession "parent" tables (#114)
The accessions table has many-to-one relationships with some secondary tables.
These can now be represented as separate namespaces and accessed using
single-value sublist fields.

This change doesn't update any of the existing fields that point to values in
those secondary tables, just adds the sublists.  A subsequent change will turn
those fields into aliases.

Also, some of the sublists have less-than-ideal names because the ideal names
are already taken. For example, there is already a `species` field for the
species name, so the species sublist is called `speciesInfo` here. A subsequent
change will introduce some aliases for the colliding names; we'll update the
client to use those new aliases, then rename the sublists.