PROOF OF CONCEPT: API to retrieve search schema
Add an API endpoint `/api/v1/search/schema` to dump out the full list of available
search tables and the relationships between them.

Current output is included in the file `search-schema.json` so it's easier to give
feedback. (That file wouldn't be included in the final version of this change.)

Don't worry too much about the implementation at this point.