TRACK-466 Handle circular references among namespaces (#120)
Previously, namespaces were constructed using an implicit tree structure, where
each one was created as part of its the initialization of its parent in the
data hierarchy.

That works when the namespaces genuinely form a tree structure, but it won't
work once they form a more complicated graph with cycles and cross-linking. For
example, `species` will appear as a sublist of `plants` as well as
`accessions`.

Fix it by switching to a two-phase initialization scheme where we create all
the namespaces and then link them together (by initializing their `sublists`
fields) only after they all exist.