TRACK-466 Add organizations, projects, sites namespaces (#121)
Allow searches to ask for information about the upper layers of the data model.

The list of fields in these namespaces is minimal for now, because we're still
finalizing requirements like how a project's location should be modeled. This is
mostly just to get the top-level structure in place so we can start expanding
the search API into other parts of the data model.