TRACK-513 Add aliases for fields that ought to be sublists (#123)
There are a few accession fields that refer to values in secondary tables and
whose field names are the names that we'd ideally want to use for the
corresponding sublists. To avoid name collisions, the sublists have `Info`
suffixes for now.

Add alternate aliases the client can use temporarily. The migration sequence
will look like this:

1. Add alternate names (this change)
2. Update the client to use the alternate names
3. Remove the offending fields and rename the sublists
4. Update the client to use the flattened sublist names instead of the aliases
5. Remove the aliases

For example, `species` is the existing field for the species name, which is
pulled from a separate table. Ideally, we'd want the separate table to
correspond to a sublist called `species`, at which point you could ask for
`species_name` to get the name. But thanks to the name collision, you
have to do `speciesInfo_name`. This diff adds an alias `speciesName`.
Once the client is using `speciesName` instead of `species`, we can
change the sublist's name to `species` which is what we want long term.