TRACK-463 Add general-purpose search endpoint (#127)
`/api/v1/search` is a general-purpose search endpoint that allows searching
arbitrary parts of the data model. It is similar to `/api/v1/seedbank/search`
except that callers supply the prefix to search.

Currently, it can be used to search organizations, projects, sites, facilities,
and accessions. Upcoming changes will add search fields for more parts of the
data model.