TRACK-467 Add layers, features, plants to search (#133)
Add the GIS-related parts of our data model to the hierarchy of search fields.

Geometry values are returned as strings whose contents are GeoJSON objects.
For now, they can only be returned as search results and used as sort keys
(though sorting by a geometric shape isn't too useful). They can't be used as
filter criteria because none of the existing filter operations really apply to
them -- exact-match might be reasonable, but coordinate system conversion makes
its semantics a bit muddled.

Plant fields are currently represented as a sublist of the features namespace.
This is somewhat at odds with our desire to move toward an API where a plant is
just a type of feature, but flattened sublists can be used to pull the plant
fields up a level so they look like they're feature fields. This will probably
need some refinement once we start using it for real.