TRACK-518 Create organization with country + state
Update the data model so clients can create organizations with the correct set of
information per the account administration designs, and add an endpoint for it.

The data model changes:

* New `countries` and `country_subdivisions` tables to hold the list of valid
  countries and (currently) US states. The latter table is set up so that in the
  future, we can include subdivisions of other countries by importing the
  ISO 3166-2 data set.
* Add `country_code` and `country_subdivision_code` columns to `organizations`
  to hold the organization's location data.
* Drop `organizations.location` since it's no longer needed.

This also addresses TRACK-492 (add new fields to organization) by returning the
new fields in the `GET /api/v1/organizations/{id}` response payload.