Fix compiler warning; treat warnings as errors
Fix a shadowed variable name that's been in the code for a month or so. To stop
that kind of thing from creeping into the code again, turn on the compiler
option that treats warnings as errors. We can revisit it if it turns out to cause
too many false positives.