TRACK-520 Track users who are invited to organizations
Add a column to `organization_users` to track whether a user has been invited to
the organization but hasn't yet accepted the invitation.

Users shouldn't have access to organizations until they accept their invitations.
Update the permission code to ignore pending invitations when it computes the list
of projects and organizations the user has permission to access.

There is no way for a client to create a user with a pending invitation yet; this
just lays the groundwork for adding an invitation API.