Fold SearchFieldNamespace into SearchTable
The search code had two metadata classes for each searchable database table.
This was because the way the application's data model was represented in the
search API didn't quite correspond 1:1 to the way the underlying database
tables were organized: some fields in child tables were represented as values
in the parent tables in order to present a flat view of the
data to the front end.

We now use flattened sublists and alias fields to bridge that gap, so there's no
longer any need to distinguish between a "search table" and a "search field
namespace."

Remove `SearchFieldNamespace` and fold its data into `SearchTable`.