TRACK-549 Utility to generate session data for Cypress tests
The recent dependency upgrades changed the format of serialized session data,
and the frontend integration test suite was still populating its database with
a fake non-expiring session in the old format.

Rather than hand-constructing a new session, add a utility class to generate
one and output its hexadecimal representation for inclusion in a database
initialization script on the front end. Add a Gradle task to run the utility.