Remove old-style filters field from new search endpoint
The accession search API payload has a `filters` field that takes a list of search
filters. This has been superseded by the more powerful `search` field that takes
an arbitrary Boolean expression. The old field is still supported for backward
compatibility, but we only need to care about that on the accession search
endpoints; there is no old code that calls the non-accession-specific search
endpoint.