TRACK-551 Return CSV error responses
If the client sends a request with an `Accept: text/csv` header and there's an
error, we need to be able to send a CSV response since that's the only kind of
response that it's valid to return.

Update our exception-handling code to generate CSV responses if needed, but to
prefer JSON if possible.

Since the only real use case for CSV error responses at the moment is the "export
to CSV" search endpoint, and that endpoint throws `BadRequestException` if there
are input validation errors, add a handler for the superclass of that exception.
All the exceptions in that hierarchy are explicit "return an HTTP error response
with status code X" errors, so we can use the status code indicated by the
exception.