TRACK-551 Support CSV export of search results
Make `/api/v1/search` return a CSV file if the incoming request specifies
`text/csv` in its `Accept` header.

Allow the client to pass a maximum result count of 0 to indicate no limit, since
pagination isn't as useful for CSV downloads as it is for interactive search in
the web UI.

The existing `/api/v1/seedbank/search/export` endpoint continues to work as
before.