Work around bogus Gradle warning message
Gradle uses an old Kotlin version internally, and that old Kotlin version
doesn't know about recent JVM releases. So the `buildSrc` build spits out a
warning about the target JVM version being different for Java and Kotlin code.

This will be fixed in the next minor Gradle release, but in the meantime, set the
target JVM version to Java 8 to stop the warning. That might cause the jOOQ code
to be compiled using slightly less efficient bytecode but the performance
difference should be infinitesimal.