TRACK-585 Add organization ID to species endpoints
Add an `organizationId` parameter to all the species-related endpoints that will
need to include it once we change the species data model to add per-organization
species lists. Depending on the endpoint, it is either a query string parameter
or a payload field.

The parameter currently does absolutely nothing. But with this change, it is now
included in the OpenAPI schema and will thus show up in autogenerated client code.

Once this is merged, the client can be updated to start including the organization
ID in all its species-related requests. After that, the parameter will become
required rather than optional.