Support partial schema docs generation
During development, it's handy to be able to look at the data model
diagram, but you usually only care about one or two of the subsets of
the tables that we generate diagrams for. Add support for generating
the docs for specific slices of the data model rather than always
generating all of them.