Use consistent Postgres Docker image for build and test (#172)
We configured the Docker image that gets used for code generation separately
from the one that gets used for running database tests, but we want the two
to always match. Move the image settings to Gradle properties and write them to
a generated source file so they're used verbatim by tests.

This also allows you to use a different image locally by overriding the
properties on the command line or in `$HOME/.gradle/gradle.properties`.