TRACK-649 Add per-user flag to projects
To support seed banks being accessible organization-wide, add a project-level
flag to control whether or not the project has per-user permissions.

There is currently no way for a client to create a non-per-user project; a later
change will use the new flag when it automatically creates seed bank projects in
new organizations.