TRACK-648 Add hidden flag to projects (#176)
We'll be automatically creating a seed bank facility when a new organization is
created, which requires a project and a site. But we don't want the project and
its children to appear in the UI. Add a project-level flag to indicate to the
front end that it should hide the project and its children.

The flag doesn't do anything on the server side; it's purely for client use.

Currently the flag is read-only; we'll set it when we auto-create the seed bank
during organization creation, but clients aren't yet able to set it.