Insert mock user into users table in tests
Tracking which users created/modified things will mean adding user ID columns to a
number of tables. Those user IDs will need to refer to IDs that exist in the users
table. That has an impact on tests, where previously we could get away with just
using an in-memory mock object. We'll still do that, but the mock needs to have a
user ID, and the user will need to exist in the users table.

Address this by introducing a helper function to create a mock user with a
reasonable set of default behaviors, and modify `insertSiteData()` to insert a
user before it inserts the organization and so forth.