TRACK-421 Support per-organization species lists (#182)
Update the data model so that each organization can have its own list of species.

From the API client's perspective, the only change here is that the response to
`/api/v1/species/name/{id}` now includes an organization ID. We already made it
mandatory for the client to pass in an organization ID where needed.

Internally, the code now makes use of the `species_options` table, which
previously existed but didn't do anything. The presence of a row in that table
is what determines whether or not a species is included in the organization's
list of species.