TRACK-678 Add system user (#184)
We're going to add "created by" and "modified by" fields to a bunch of entities to
keep track of which users are making changes. But some changes are made
automatically by the system, e.g., advancing accessions to new states based on the
passage of time. In those cases, the changes need to be attributed somehow, but
there's no individual user to attribute them to.

Add a special "system" user that we can use for automated changes that aren't
initiated by a specific individual user.

The system user needs to be able to modify data across all organizations and
projects. Rather than explicitly adding it as a member of every project, instead
we hardwire its permission checks to always return `true`.

There's no clean way to hardwire permission checks in `UserModel` without
introducing a bunch of special-case code that could potentially be run by mistake
for regular users; it's simpler and less bug-prone to introduce a separate model
class that can only ever represent the system user, and allow it to be used in
place of `UserModel`.

To that end, this change:

* Adds a `TerrawareUser` interface that declares all the methods and properties
  that aren't specific to individual (human or API client) users. `currentUser()`
  is defined as returning a `TerrawareUser` implementation, and most code doesn't
  need to care whether that's the system user or a regular user.
* Adds a `SystemUser` class that implements `TerrawareUser` and has hardwired
  permission checks. This is a singleton: there is exactly one system user, and
  it is impossible to log in as the system user.
* Renames `UserModel` to `IndividualUser` so the class names are consistent.
* Updates `DailyTaskRunner` so that scheduled tasks are run as the system user.