TRACK-659 Add users to search API
There is now a `users` sublist under both organizations and projects that can be
used to search for user information.

These sublists are actually information about the users' memberships in the
organization or project; each of them has a `user` single-value sublist that
contains the actual user information.

Search results will only include users that are members of at least one of the
current user's organizations.

Similarly, the user information has `organizations` and `projects` sublists, which
will only include organizations and projects the current user has permission to
see.