TRACK-594 Add created-by and modified-by columns to tables (#185)
Add `created_by` and `modified_by` columns to all the tables that represent
things that can be directly edited via the API. That includes most tables, but
excludes things like withdrawals which can only be edited by editing the
accession.

Add `created_time` and or `modified_time` columns to tables that didn't already
have them.

These new columns are all non-nullable. The database migration attempts to
populate them with reasonable defaults, which in most cases means the owner of
the organization they're associated with.

This change touches a lot of files, but most of the changes are pretty similar:
populate the new columns in the create/update methods, and modify tests to
expect values in the new columns.