Upgrade to jOOQ 3.16
The latest version of jOOQ has started adding support for spatial data types
such as the PostgreSQL `GEOMETRY` type. For now, we want to continue using the
PostGIS library since it is more feature-rich, but the binding code that
converts raw database values to classes from the PostGIS library needs to be
updated to recognize that the "raw database values" are now jOOQ `Geometry`
objects rather than simple strings.

This change doesn't affect the application code aside from `GeometryBinding`,
but it's required if we ever want to upgrade to subsequent jOOQ versions to get
additional features.