TRACK-745 Require names to be nonempty
Previously, the server accepted requests to create or update organizations,
projects, and sites with zero-length names. From the user's point of view, we were
effectively not treating names as required.

Add validation annotations to the appropriate request payloads, and add a handler
to render the validation failure exceptions in human-readable form. Now if you
pass in an empty name, you'll get back an HTTP 400 with a standard JSON error payload:

    {
        "error": {
            "message": "Field value invalid: name must not be empty"
        },
        "status": "error"
    }