TRACK-780 Add storage location management to admin UI (#194)
Currently, the only way to configure the list of storage locations that show up
in the seed bank UI is to connect to the database and run SQL commands by hand.

Since we're in the middle of rethinking how we model facilities and such, it
doesn't make sense to add storage location management to the frontend. But as a
stopgap, we at least need a way to set things up for customers.

Add a section to the facility page in the admin UI where we can manage the list
of storage locations for a facility.