TRACK-774 Create default storage locations (#196)
We create a default seed bank facility for new organizations, but if users try to
use it, the list of storage locations is empty. Eventually we'll probably need to
give users more control over this, but for now, create three refrigerators and
three freezers as part of the default setup of a seed bank, which is the correct
setup for our containerized seed banks.