TRACK-775 Reject negative seed quantities
Previously, there was no sanity checking on the quantities of withdrawals or
germination tests and it was possible to withdraw 1000 seeds from an accession
with 50 seeds remaining. Add validation checks to reject updates that cause the
remaining quantity to drop below zero.

It was also possible to specify a zero or negative total accession size; catch
that as well.