TRACK-838 Send email when device manager goes idle (#210)
Add logic to send an email alert to the facility's alert recipients when too much
time has passed since the most recent batch of metrics for the facility's devices.

This works by keeping track of the most recent data from each facility. There
is a facility-level setting (available via the admin UI) to control how much
time is allowed to pass between updates from the device manager before a
facility is considered idle. When more than that amount of time has passed, the
facility is marked as idle and email is sent.