TRACK-654 Remove obsolete filters field from search API (#212)
The initial version of the search API had a less sophisticated way to specify
search criteria. We kept it around for backward compatibility since it was
still being used in the web app. The last use of it has now been removed from
the frontend code: it was only being used by the accession number search bar,
which was removed from the UI. Get rid of it on the server side.