TRACK-843 Include organization ID in email notifications
The link in an "added to organization" email notification should take the user
directly to the organization in question.

Since this won't be the last time we need to construct a link to a specific place
in the web app, add a helper class to consolidate all the link construction logic.