Renovate: Adjust timing of updates
For patch releases, scan for updates on weekends so we don't get them sprinkled
piecemeal during the week.

For major or minor releases, don't create a PR until the release has been out for
at least 3 days. That way we can pick up any patch releases that come right on the
heels of a significant change.